+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,40 @@
 
 ### Patch Changes
 
+- Updated dependencies [dee8b05]
+  - ai@6.0.0-beta.29
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [521c537]
+  - ai@5.1.0-beta.28
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.27
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [c99da05]
+  - ai@5.1.0-beta.26
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [457f1c6]
+  - ai@5.1.0-beta.25
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [90e5bdd]
   - ai@5.1.0-beta.24