+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @ai-sdk/xai
 
+## 3.0.0-beta.11
+
+### Major Changes
+
+- dee8b05: ai SDK 6 beta
+
+### Patch Changes
+
+- Updated dependencies [dee8b05]
+  - @ai-sdk/openai-compatible@2.0.0-beta.10
+  - @ai-sdk/provider@3.0.0-beta.6
+  - @ai-sdk/provider-utils@4.0.0-beta.10
+
+## 2.1.0-beta.10
+
+### Patch Changes
+
+- Updated dependencies [521c537]
+  - @ai-sdk/provider-utils@3.1.0-beta.9
+  - @ai-sdk/openai-compatible@1.1.0-beta.9
+
+## 2.1.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies [e06565c]
+  - @ai-sdk/provider-utils@3.1.0-beta.8
+  - @ai-sdk/openai-compatible@1.1.0-beta.8
+
 ## 2.1.0-beta.8
 
 ### Patch Changes